aws code deploy를 위해 코드 수정

+++ scripts/remove_server.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-pwd
-docker-compose down
+if docker ps ; then
+    docker-compose down 
+    docker rm -fv test-helper_ai_small
+fi
